kubeflow_women_clothing_ecom_reviews/pipeline/components/feature_engineering/requirements.txt
pyyaml==5.4.1
pandas==1.1.5
textblob==0.15.3
scikit-learn==0.24.1
# for local testing
# kfp==1.4.0